refactor: update timestamp parser test to be compatible with Go 1.23 (#37)
In Go 1.23, `time.Parse` was changed as follows:
> Parse and ParseInLocation now return an error if the time zone offset
is out of range.

Previously, it would accept an offset like +99:00. Therefore, the
optimized parser in this evaluation
module also modeled that behavior. 

One path forward would be to update the parser to match `time.Parse`'s
behavior, but that would be a breaking
behavioral change with unknown ramifications (and might require a major
version of the eval module.)

For now, mark the test as nonstandard to check that it is parseable by
the optimized parser, but not necessarily
that it matches stdlib behavior.